# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ecommerce API in go using Gin frame work",
        "title": "E Commerce API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/bestselling": {
            "get": {
                "description": "Fetch  the best selling products from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Best selling products",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Type of search: 'product' or 'category'",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User was deleted",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
        "/admin/login": {
            "post": {
                "description": "Authenticate admin credentials and generate JWT token for authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Admin Login",
                "parameters": [
                    {
                        "description": "Credentials for authentication ( username \u0026 password)",
                        "name": "Credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.adminDetail"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successfully logged",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "401": {
                        "description": "Invalid username or password",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "501": {
                        "description": "Error binding data",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
        "/admin/signup": {
            "post": {
                "description": "Authenticated admin can create new user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Signup"
                ],
                "summary": "Admin SignUp",
                "parameters": [
                    {
                        "description": "Create Admin",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.adminDetail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "New admin created",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/admin/user": {
            "get": {
                "description": "get list of all registered admins",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Users"
                ],
                "summary": "list of users",
                "responses": {
                    "200": {
                        "description": "List of users",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controller.UpdateUserData"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch user data",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/admin/user/{id}": {
            "delete": {
                "description": "Delete an existing user from admin side",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Users"
                ],
                "summary": "Delete user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "500": {
                        "description": "Failed to delete user",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            },
            "patch": {
                "description": "Edit user details based on user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Users"
                ],
                "summary": "Edit user details",
                "parameters": [
                    {
                        "description": "Update user info",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.UpdateUserData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "406": {
                        "description": "Failed to bind data",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "500": {
                        "description": "Failed to update details",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
        "/admin/userblock/{id}": {
            "patch": {
                "description": "Update User Bloking status as Blocked or Unblocked",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin/Users"
                ],
                "summary": "Block user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User blocked or unblocked successfully",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "json"
                        }
                    },
                    "500": {
                        "description": "Server error while trying to change blocking status",
                        "schema": {
                            "type": "json"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.UpdateUserData": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "controller.adminDetail": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}